third_party/blink/web_tests/platform/generic/storage/indexeddb/readonly-properties-expected.txt

Test IndexedDB readonly properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "readonly-properties.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('foo');
trying to set readonly property objectStore.transaction
objectStore.transaction = this
PASS objectStore.transaction is still [object IDBTransaction]
PASS successfullyParsed is true

TEST COMPLETE
